+++ packages/babel-plugin-pdsl/src/literals.test.js
@@ -5,7 +5,7 @@ const {
   numericLiteral,
   stringLiteral
 } = require("./literals");
-const { grammar, tokens } = require("pdsl/grammar");
+const { grammar, tokens } = require("pdsl/index/grammar");
 const generate = require("@babel/generator").default;
 
 describe("numericLiteral", () => {
