+++ packages/pdsl/helpers/wildcard.js
@@ -1,0 +1,5 @@
+module.exports.createWildcard = () =>
+  function wilcard() {
+    // never going to fail so no need to do error reporting
+    return true;
+  };
