+++ packages/pdsl/index/generator.js
@@ -1,4 +1,4 @@
-const { getRawHelpers } = require("./helpers");
+const { getRawHelpers } = require("../helpers");
 const { val } = getRawHelpers();
 const {
   isPredicateLookup,
