+++ packages/pdsl/index/grammar.js
@@ -1,4 +1,4 @@
-const { getRawHelpers } = require("./helpers");
+const { getRawHelpers } = require("../helpers");
 const {
   and,
   arrArgMatch,
