+++ packages/pdsl/package.json
@@ -2,8 +2,8 @@
   "name": "pdsl",
   "version": "5.2.5",
   "description": "Predicate DSL",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "index/index.js",
+  "types": "index/index.d.ts",
   "author": "Rudi Yardley",
   "license": "MIT",
   "repository": "github:ryardley/pdsl",
